Add additional validation to R1CS proof points.
This follows the same logic as in the rangeproof case: we don't have an a
priori reason to think that there's a security problem with allowing these
points to be the identity, but this is an area where people make mistakes in
under-specification of assumptions, etc., and there is no valid reason for
these points to be the identity, so this works as a defense-in-depth mechanism.

The extra validation is not applied to the value commitments, in order to allow
commitments to the zero value with zero blinding factor.